Update to `xtask` cli to accomodate runner commands
Runners for a few tutorials will just have options that are passed to
the program being run (the tutorial runner) or may have some commands
each individually taking options. This was not handled initially as all
the runners were run like the former. Now all the options and arguments
are passed to the runner as `last` arguments.